Add Close() method to Project to release resources
Problem Observed

Goroutine leaks because there is no way to "close" projects

Problem Detail

NewProject() creates a defaultListener in it.
The listener starts a new goroutine in NewDefaultListener().
We currently don't have a method to stop this goroutine.

Suggested Resolution

This commit adds a new method Close() to release resources
tied to a Project.

Signed-off-by: Iwasaki Yudai <yudai@arielworks.com>